Module Murder_generator.Attribute

module type Attribute = sig ... end

A signature for the various attribute/constructor structures. This signature is used in particular for the player attributes and the contact attributes.

module PlayerAttribute : Attribute

A module to express attributes and constructors for players and objects.

module ContactAttribute : Attribute

A module to express attributes and constructors for contacts between players.

type constructor_maps = {

player : PlayerAttribute.constructor_map;

contact : ContactAttribute.constructor_map;

}

This data structure stores all the informations about constructors.

val empty_constructor_maps : constructor_maps

An empty structure.

val object_type : PlayerAttribute.attribute

A special internal attribute for objects.

type attribute =

| PlayerAttribute of PlayerAttribute.attribute

| ContactAttribute of ContactAttribute.attribute

A generic attribute type for modules who don’t need to precisely understand how attributes work, merging both kinds of attributes.

type constructor =

| PlayerConstructor of PlayerAttribute.constructor

| ContactConstructor of ContactAttribute.constructor

Similarly, a generic constructor type.