Module Murder_generator.Driver

val parse_lexbuf : string -> Stdlib.Lexing.lexbuf -> Ast.declaration list

From the file name and the lexing buffer, calls the parser.

val parse_relation : string -> Relation.t

Parses a relation obtained through Relation.to_string.

type intermediary

A type to store intermediate informations about parsing.

exception UnexpectedCommandInBlock of string * string

States that the block named as the first string has an invalid command in its block, which is described by the second string.

exception DefinedTwice of string * string * string

An error indicating that something of kind described by the first string has been defined twice, the second string being its name, and the third describe its enclosing block.

exception Undeclared of string * string * string

An error indicating that something of kind described by the first string has been used without being declared, the second string being its name, and the third describe its enclosing block.

exception CircularDependency of string * string

An error indicating that there is a loop definitions. The first argument is the definition kind, the second its name.

exception SelfRelation of string * string

An error indicating that a player has been associated a relation to itself in the given element.

exception TranslationError of string * string * Ast.translation

An unexpected translation item has been given in an element of the kind described by the first argument, whose name is the second argument.

exception VacuumElement of string

The given element can’t be applied (typically because some constraints are incompatible).

exception UnsatisfyableEventSequence of string

If a sequence of event is not satisfyable in an element, report the element. A sequence of event may not be satisfyable if for instance an event prevents an event of a given kind to appear after it, but that this precise one actually appears in the event.

val prepare_declarations : intermediary -> Ast.declaration list -> intermediary

Treating the data sent from the parser is done in two phases. The first phase is to assimilate each globally declared names into identifiers. It expects the previously computed intermediary type. This function might throw UnexpectedCommandInBlock or DefinedTwice.

val empty_intermediary : intermediary

An initial empty intermediary element.

val is_intermediary_final : intermediary -> bool

States whether the intermediary type is waiting for other definitions to be complete. This function is usually called once each file have been loaded and merged into an intermediate: if this function then returns false, an error has been made somewhere.

val categories_to_be_defined : intermediary -> string Libutils.PSet.t

Return a set of missing categories.

val events_to_be_defined : intermediary -> string Libutils.PSet.t

Return a set of missing events.

val attributes_to_be_defined : intermediary -> string Libutils.PSet.t * string Libutils.PSet.t

Return a set of missing attributes and contacts.

val constructors_to_be_defined : intermediary -> (string * string) Libutils.PSet.t * (string * string) Libutils.PSet.t

Return a set of missing constructors for attributes and contacts. The first element of each couple is the attribute name, then the constructor name.

val tags_to_be_defined : intermediary -> (string * string) Libutils.PSet.t

Return a set of missing tags. The first element of each couple is the language code, then the tag.

type state

This type stores all the informations returned by parse.

val parse : intermediary -> state

Once each file have been parsed and put into a single intermediary result, one can treat the full data and produce a final state. This function might throw any of the exceptions defined in this file.

val elements : state -> (Libutils.Id.t, Element.t) PMap.t

This map stores the identifiers of each element.

val get_element_name : state -> Libutils.Id.t -> string option

Given an element identifier, returns its name.

val get_all_elements : state -> Translation.language -> Libutils.Id.t Libutils.PSet.t -> int -> Libutils.Id.t list

Return all defined elements in this state that are compatible with the provided language, the chosen set of categories (given as identifiers) and at most this number of players.

val all_categories : state -> Libutils.Id.t list

Once the state has been built, categories can be listed.

val total_number_of_elements : state -> int

Returns the number of elements defined in the state.

val number_of_elements : state -> Translation.language -> int

Returns the number of elements available for a particular language.

val get_category_dependencies : state -> Libutils.Id.t -> Libutils.Id.t Libutils.PSet.t

Takes a category identifier and returns the list of categories that this category depends on.

val get_attribute_dependencies : state -> Attribute.attribute -> Libutils.Id.t Libutils.PSet.t

Takes an attribute and returns the list of categories that this attribute depends on.

val get_constructor_dependencies : state -> Attribute.constructor -> Libutils.Id.t Libutils.PSet.t

Takes a constructor and returns the list of categories that this constructor depends on.

val get_element_dependencies : state -> Libutils.Id.t -> Libutils.Id.t Libutils.PSet.t

Takes an element identifier and returns the list of categories that this element depends on.

val get_translations : state -> Translation.element

Return the translations for elements.

val get_constructor_maps : state -> Attribute.constructor_maps

Return the constructor maps generated while loading the files.

type import_information = {

constructor_maps : Attribute.constructor_maps;
Information about constructors.

event_id : (string, Libutils.Id.t) PMap.t;
A unique identifier for each event.

event_informations : (Libutils.Id.t, bool * bool * int Events.translation) PMap.t;
Information for each event, organised over event names. In particular:

whether it is a phantom event,

whether it is a blocking event,

how it is translated.

event_kinds : (Libutils.Id.t, (int, int Events.kind Libutils.PSet.t) PMap.t) PMap.t;
The kind associated to each event, organised over event names.

}

Necessary information to import a state.

val get_import_information : state -> import_information

Return information for importation.