Module type InOut.T

This is the signature specified in this file. It is satisfied by the various files lib/inOut_*.ml.

val pause : unit -> unit Lwt.t

Pause the program for a short amount of time.

val stopLoading : unit -> unit Lwt.t

Stop the loading animation.

val startLoading : unit -> unit Lwt.t

Start the loading animation.

val setLoading : float -> unit Lwt.t

Given a float between 0. and 1., set the corresponding loading in the loading animation. This function does not change the state of the animation: if the animation is stopped, it will not start it.

val set_printing_mode : unit -> unit

Set or unset the printing mode, a mode more adapted to printing.

val unset_printing_mode : unit -> unit

val log : string -> unit

Log the given message.

val get_file : string -> string Lwt.t

Fetch a file from an address and returns its content.

val get_parameters : unit -> (string * string) list

Get parameters from the current address, in the form of a list of string.

val set_parameters : (string * string) list -> unit

Write parameters to the address.

val languages : string list

The local set of accepted languages.

type node

An abstract type for representing nodes.

type ('a, 'b) interaction = {

node : node;
The node itself

get : unit -> 'b;
Getting its value

set : 'a -> unit;
Setting its value

onChange : ('a -> unit) -> unit;
Calling a callback each time the value is changed

lock : unit -> unit;
Lock the node: no one can change its value

unlock : unit -> unit;
Unlock the node

locked : unit -> bool;
Current lock status

onLockChange : (bool -> unit) -> unit;
Calling a callback each time the node is locked or unlocked.

}

A type for node interaction.

type 'a sinteraction = ('a, 'a) interaction

The type of safe interactions.

val synchronise : 'a sinteraction -> 'a sinteraction -> unit

Making two safe interactions of the same time be copies of one another.

val block_node : node block -> node

Converts the block to a node.

val print_node : ?⁠error:bool -> node -> unit

Adds the node to the response div in the main webpage. If error is true, the node is highlighted as an error.

val print_block : ?⁠error:bool -> node block -> unit

A composition of add_spaces, block_node, and print_node.

val clear_response : unit -> unit

Clears the response div in the main webpage.

val createNumberOutput : int -> node * (int -> unit)

Create a text output as a number which can be later reset.

val createTextOutput : string -> node * (string -> unit)

Create a text output as a string which can be later reset.

val createNumberInput : ?⁠min:int -> ?⁠max:int -> int -> int sinteraction

Create a (positive) number input with default value given as argument.

val createTextInput : string -> string sinteraction

Create a text input with default value given as argument.

val createListInput : (string * 'a) list -> (string, (string * 'a) option) interaction

Create a drop-down list where the user can choose one of its items.

val synchroniseListInput : (string, (string * 'a) option) interaction -> (string, (string * 'a) option) interaction -> unit

Synchronise two drop-down lists.

val createResponsiveListInput : (string * 'a) list -> string -> (string -> (string * 'a) list) -> (string * 'a) list sinteraction

Create a text input meant to return a list of things. Each time that the user type a string, it is fed to its argument function. The strings of the returned list is shown to the user. If the user chooses one of these elements, it is added to a list displayed next to the input. This final list can be fetched with the function returned with the node. The initial list and a placeholder string is given to the function to help the user.

val createPercentageInput : float -> float sinteraction

Create a range input between 0. and 1..

val createDateInput : Libutils.Date.t -> Libutils.Date.t sinteraction

Create a date input.

val createSwitch : string -> string option -> string option -> string option -> bool -> bool sinteraction

Create a switch button. The first string is the text associated with the button, to which can be added three facultative texts: one serving as a description, one added afterwards for when the button is on, and one when the button is off.

val createFileImport : string list -> (unit -> unit Lwt.t) -> node * (unit -> (string * string) Lwt.t)

Create a button to import a file. It takes as argument the list of extension it accepts. It also takes as argument a function called before loading the file into memory. As of the other reading functions, it returns the created element and a reading function. This reading function returns both the file name and its content.