Module Main.Main

This entire file is parameterised by an interface as specified in the InOut Module.

Parameters

IO : InOut.T

Signature

val webpage_link : string

val webpage_issues : string

val trace : string list Stdlib.ref

A trace of the menu function to help debugging.

val add_trace : string -> unit

Adding a message in the trace

val get_trace : 'a -> string list

Get the full trace.

val errorTranslationsDefault : string * string * string * string

The default error messages.

val errorTranslations : (string * string * string * string) Stdlib.ref

The translations needed to print error messages.

val file_signature : 'a -> string

val get_translations : 'a -> (string Translation.t * Translation.language list) Lwt.t

Getting and parsing the translations file.

val print_list : string -> string list -> string

Prints a list of strings, andw being the word for “and” in the current language. This function makes use of the Oxford comma.

val get_data : 'a -> Driver.state Lwt.t

Get and parse each data file.

val get_names : Driver.state Lwt.t -> Names.t list Lwt.t

Get and parse each name file.

type player_information = {

name : string;
Character name

complexity : int;
Complexity

difficulty : int;
Difficulty

attributes : Attribute.PlayerAttribute.constructor list;
Preset attributes

contacts : Attribute.ContactAttribute.constructor list list;
Preset contacts

}

A type to store player information

type parameters = {

language : Translation.language option;

player_number : int;

general_level : float;

general_complexity : float;

play_date : Libutils.Date.t;

computation_power : float;

categories : Libutils.Id.t Libutils.PSet.t option;

player_information : player_information list;

chosen_productions : string Libutils.PSet.t;
The name of each chosen production.

}

A type to store what each page of the menu provides.

val urltag_lang : string

URL tags

val urltag_number : string

val urltag_level : string

val urltag_complexity : string

val urltag_date : string

val urltag_power : string

val urltag_categories : string

exception InvalidUrlArgument

val create_player_information : Names.t list -> 'a -> parameters -> player_information list

Update player informations from the parameters such that it matches the player_number data.

val get_categories : Driver.state -> parameters -> Libutils.Id.t Libutils.PSet.t

Get the categories from the parameters, but return the default value if not defined.

val get_all_elements : Driver.state -> parameters -> Libutils.Id.t list

Get all elements corresponding to these settings.

val attribute_functions : (Attribute.PlayerAttribute.constructor_map -> Attribute.PlayerAttribute.constructor list) * (Attribute.PlayerAttribute.constructor_map -> Attribute.PlayerAttribute.constructor -> Attribute.PlayerAttribute.attribute option) * (Attribute.PlayerAttribute.constructor_map -> Attribute.PlayerAttribute.attribute -> string option) * (Attribute.PlayerAttribute.constructor_map -> Attribute.PlayerAttribute.constructor -> string option) * (Attribute.PlayerAttribute.constructor_map -> Attribute.PlayerAttribute.attribute -> Attribute.PlayerAttribute.constructor -> bool) * (Attribute.constructor_maps -> Attribute.PlayerAttribute.constructor_map) * (Attribute.PlayerAttribute.attribute -> Attribute.attribute) * (Attribute.PlayerAttribute.constructor -> Attribute.constructor)

In order to factorise between attributes and contact, the following two declarations defines some functions in common between the two structures, so that subfunctions can be reused for both.

val contact_functions : (Attribute.ContactAttribute.constructor_map -> Attribute.ContactAttribute.constructor list) * (Attribute.ContactAttribute.constructor_map -> Attribute.ContactAttribute.constructor -> Attribute.ContactAttribute.attribute option) * (Attribute.ContactAttribute.constructor_map -> Attribute.ContactAttribute.attribute -> string option) * (Attribute.ContactAttribute.constructor_map -> Attribute.ContactAttribute.constructor -> string option) * (Attribute.ContactAttribute.constructor_map -> Attribute.ContactAttribute.attribute -> Attribute.ContactAttribute.constructor -> bool) * (Attribute.constructor_maps -> Attribute.ContactAttribute.constructor_map) * (Attribute.ContactAttribute.attribute -> Attribute.attribute) * (Attribute.ContactAttribute.constructor -> Attribute.constructor)

val main : unit Lwt.t

The main script.