Module Murder_generator__Attribute.PlayerAttribute

A module to express attributes and constructors for players and objects.

type attribute

The type of attributes.

type constructor

The type of constructors.

type constructor_map

Values are just constructor identifiers. Each attribute is associated with a given set of possible constructors (which are really just names). The following table keeps track of the constructor names.

val empty_constructor_map : constructor_map

An empty constructor map.

val attribute_name : constructor_map -> attribute -> string option

Return the name of an attribute.

val constructor_name : constructor_map -> constructor -> string option

Return the name of a constructor.

val constructor_attribute : constructor_map -> constructor -> attribute option

Return the associated attribute of a constructor.

val constructors : constructor_map -> attribute -> constructor list option

Return the list of constructors associated to an attribute.

val all_constructors : constructor_map -> constructor list

Returns all defined constructors.

val declare_attribute : constructor_map -> string -> bool -> attribute * constructor_map

Declare an attribute, returning its associated normal identifier. The string is the name of the attribute and the boolean states whether it is internal. If already declared, its previously-set identifier is still returned, but its internal status is updated.

val declare_constructor : constructor_map -> attribute -> string -> bool -> constructor * constructor_map

Declare a new constructor for an attribute. The string is the name of the constructor and the boolean states whether it is internal. If already declared, its previously-set identifier is still returned, but its internal status is updated.

val get_attribute : constructor_map -> string -> attribute option

Get the attribute identifier from its name.

val get_constructor : constructor_map -> attribute -> string -> constructor option

Get the constructor identifier from its attribute and its name.

val remove_constructor : constructor_map -> constructor -> constructor_map

Users can remove categories before the story generation. This function removes a constructor, probably because it was associated to an unwanted category.

val declare_compatibility : constructor_map -> attribute -> constructor -> constructor -> constructor_map

State that the first constructor is compatible with the second.

val is_compatible : constructor_map -> attribute -> constructor -> constructor -> bool

State whether the first constructor is compatible with the second.

val is_internal : constructor_map -> attribute -> constructor -> bool

State whether the given constructor is internal.