Module Murder_generator__History

Module History Stores the timeline of characters.

type character = Libutils.Id.t

type event = {

event_begin : Libutils.Date.t;
Beginning of the event

event_end : Libutils.Date.t;
End of the event

event : character Murder_generator.Events.t;
The actual event

}

This type enriches Events.t with a date and duration.

val generate_event : Libutils.Date.t -> character Murder_generator.Events.t -> event

A smart constructor for events. It takes its starting date and an event and associates it with an actual duration.

val generate_event_inv : Libutils.Date.t -> character Murder_generator.Events.t -> event

Same as generate_event, but takes as argument the ending date of an event.

val get_event_type : Libutils.Date.t -> Libutils.Date.t -> Murder_generator.Events.event_type

Infer the kind of event from its starting and ending dates.

val compatible_events : event -> event -> bool

States whether two events are compatible, that is that they do not overlap, or that they are of different types.

type t

The global timeline, storing all events that happenned. This typed is used when looking for solution to constraints and only stores the relevant interactions between events.

type final = event list

The final type, where each event have actually been assigned a particular moment in type. The events are guaranteed to be ordered along their start time.

type status =

| Normal
Any given player can only have this element once.

| Duplicable
This element can be applied without any restriction.

| Unique
This element can be applied at most once in the whole scenario.

Describes the status of elements with respect to their behaviours with timelines. Most elements are Normal, but some have special behaviours.

val compatible_and_progress : t -> character Murder_generator.Events.t -> bool Stdlib.Lazy.t option

States whether an event is compatible with a timeline. Its return value is the same as Element.compatible_and_progress: None if incompatible, Some false is compatible but does not help with the already present constraints, and finally Some true means that this event not only applies, but it helps the already present events.

val lcompatible_and_progress : t -> character Murder_generator.Events.t list -> bool Stdlib.Lazy.t option

States whether a list of events is compatible with a timeline, assuming that each event has to be in this particular order.

val apply : t -> status -> character Murder_generator.Events.t -> t

Apply an event to a timeline. This function should only be called on events for which compatible_and_progress returns Some.

val lapply : t -> status -> character Murder_generator.Events.t list -> t

Same as apply, but for a list of events for which lcompatible_and_progress returns Some. Their relative order will be conserved.

val copy : t -> t

(Deeply) copies the state.

val create_state : int -> t

Creates an empty history state for the given number n of characters, each indexed from 0 to n - 1.

val finalise : t -> Libutils.Date.t -> final

Provide an actual instantiation of each event’s date and time. It needs the date of the played scenario.

val unfinalise : final -> t

Inverse of finalise. There may be additional connections in the generated timeline than in the original one. This function only recovers what is possible to recover from the type final.

val fold_graph : ('a -> character Murder_generator.Events.t -> Libutils.Id.t -> (Libutils.Id.t list * Libutils.Id.t list) -> 'a) -> 'a -> t -> 'a

Fold over the event graph. At each step, we are given an event, its identifier, and two lists of identifiers, corresponding to the events before and after the current event.

val all_players_length : int -> character list

Given a number of player, return the associated characters identifiers. The existence of this function implies that character identifiers are fixed accross all executions and really only depends on the total number of players.