Module Murder_generator__Main

Module Main. This file is the one compiled to JavaScript, then fetched and executed to run the whole program.

module Main : functor (IO : Murder_generator.InOut.T) -> sig ... end

This entire file is parameterised by an interface as specified in the InOut Module.