Module Murder_generator__Names

Module Names This module deals with name generation.

type t

The type of name generator.

type property = Murder_generator.Attribute.PlayerAttribute.constructor

Name generator may also generate properties, that is, attribute constructors associated to the name.

val empty : t

A dummy generator that always generate the empty name.

val import : Murder_generator.Attribute.PlayerAttribute.constructor_map -> string -> t

Given the content of a name file, returns a name generator.

val generate : t -> property Libutils.PSet.t -> string * property list

Generate a random name. It takes as an argument a set of properties that the name has to be compatible with and returns a list of properties that the name additionally fulfill.

val translate : t -> unit Murder_generator.Translation.t

Translate the generator.

val is_default : t -> Murder_generator.Translation.language -> bool

State whether this generator is marked as default for this particular language.